fix: rename LangChainHelper, remove LangChainProvider, fix LDMessage[] input handling
